Add files via upload
The aim of this project was to investigate the Walmart sales data and explore top performing batches and products, sales trends of different products as well as customer behaviour and gain insight to understand the different factors that affect sales of the different branches. 

The Dataset attached is derived from Kaggle Walmart Recruiting - Store Sales Forecasting | Kaggle.